+++ R/calibrate.R
@@ -718,7 +718,7 @@ calibratable <- function(y, er, lab=c(), cc=1, BCAD=FALSE, postbomb=FALSE, bomba
 	if(one.cc)  
       this.cal <- caldist(y[i], er[i], cc=cc, BCAD=BCAD, thiscurve=this.cc, 
         deltaR=deltaR[i], deltaSTD=deltaSTD[i]) else
-          this.cal <- caldist(y[i], er[i], cc=thiscc[i], BCAD=BCAD, postbomb=postbomb, bombalert=bombalert,
+          this.cal <- caldist(y[i], er[i], cc=this.cc[i], BCAD=BCAD, postbomb=postbomb, bombalert=bombalert,
             glue=glue, thiscurve=thiscurve, cc.dir=cc.dir, is.F=is.F, is.pMC=is.pMC,
             deltaR=deltaR[i], deltaSTD=deltaSTD[i])
     this.cal <- this.cal[order(this.cal[,1]), ]